rag/requirements.txt
llama-index==0.9.15
llama-index-vector-stores-chroma==0.1.6
llama-index-embeddings-huggingface==0.1.5
sentence-transformers==2.2.2
chromadb==0.4.18
python-dotenv==1.0.0
markdown==3.5
beautifulsoup4==4.12.2
pydantic==2.5.2
fastapi==0.104.1
uvicorn==0.23.2
transformers==4.35.2
torch==2.1.0
tqdm==4.66.1
